+++ CHANGELOG.md
@@ -8,7 +8,7 @@ https://keepachangelog.com/en/1.0.0/
 - `StringPlaceholdersInWrongOrder`: Code fix generates correct code when there are more than 10 placeholders in the format string
 - `UnboundedStackalloc`: Does not trigger when using pointers, e.g. `int*`
 - `ThreadSleepInAsyncMethod`: Fixed a rare scenario in which an empty code fix would be offered
-- `ElementaryMethodsOfTypeInCollectionNotOverridden` and `StringPlaceholdersInWrongOrder` tests use `VerifyCS`
+- `ElementaryMethodsOfTypeInCollectionNotOverridden`, `LoopedRandomInstantiation` and `StringPlaceholdersInWrongOrder` tests use `VerifyCS`
 
 ## [1.21.1] - 2023-01-22
 - `StringConcatenatedInLoop`: No longer triggers for regular assignments